gemma-backend/pyproject.toml
[project]
name = "gemma-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.9.0",
    "fastapi>=0.116.1",
    "fastapi-cors>=0.0.6",
    "httpx>=0.28.1",
    "huggingface-hub>=0.34.3",
    "mlx>=0.27.1",
    "mlx-vlm>=0.3.2",
    "numpy>=2.0.0,<2.3.0",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "qrcode[pil]>=8.2",
    "requests>=2.32.4",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.54.1",
    "uvicorn>=0.35.0",
]
